refactor(registration-react): Rename function-component of FormCatalog to FormCatalogPure and final wrapped component to FormCatalog
