+++ applications/registration-react/src/pages/dataset-list-page/form-catalog/form-catalog.scss
